mysite/polls/serializers.py
from polls.models import Question
from rest_framework import serializers


class QuestionSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Question
        fields = ("id", "question_text", "pub_date")
