requirements.txt
asgiref==3.5.0
Django==3.2
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
